src/requirements.txt
ragas
datasets
ipykernel
langchain-ollama
torch
requests
sentence_transformers
transformers 
langchain
fastapi
uvicorn
pypdf
PyPDF2
jinja2
qdrant-client
ctransformers[cuda]
python-multipart
aiofiles
urllib3==1.26.18
langchain-text-splitters
langchain-experimental
langchain_qdrant
langchain_huggingface
langchain-community
huggingface-hub
qdrant-client
python-dotenv